services/backup/lc-beaty-service/requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-multipart==0.0.6

# LangChain Core (버전 유연하게)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# OpenAI
openai==1.54.3

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
